+++ src/pages/MyPage/Login.scss
@@ -1,0 +1,57 @@
+// src/pages/Home/Login.scss
+.login-form {
+    width: 100%;
+    max-width: 375px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: transparent;
+    padding: 0 43px;
+  
+    h2 {
+      font-size: 24px;
+      color: #ffffff;
+      margin-bottom: 10px;
+    }
+  
+    input {
+      width: 100%;
+      padding: 8px;
+      margin: 4px 0;
+      border: none;
+      border-bottom: 1px solid #ffffff;
+      background-color: transparent;
+      color: #ffffff;
+      font-size: 16px;
+  
+      &::placeholder {
+        color: #ffffff;
+      }
+    }
+  
+    button {
+      width: 100%;
+      padding: 12px;
+      margin-top: 20px;
+      border-radius: 24px;
+      border: none;
+      background-color: #f5c265;
+      color: #5e3a25;
+      font-size: 18px;
+      font-weight: bold;
+      cursor: pointer;
+    }
+  
+    .link-text {
+      margin-top: 16px;
+      font-size: 14px;
+      color: #ffffff;
+  
+      span {
+        color: #f5c265;
+        cursor: pointer;
+        text-decoration: underline;
+      }
+    }
+  }
+
