+++ src/pages/MyPage/MyPage.jsx
@@ -1,17 +1,13 @@
 // src/pages/Home/MyPage.jsx
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import Login from './Login';
+import SignUp from './SignUp';
+import ResetPassword from './ResetPassword';
 import './MyPage.scss';
 
 const MyPage = () => {
   const [view, setView] = useState('landing'); // landing, login, signup, success, resetPassword
-  const [userId, setUserId] = useState(''); // Store user ID input
-  const [password, setPassword] = useState(''); // Store password input
-  const [confirmPassword, setConfirmPassword] = useState(''); // Store password confirmation input
-  const [nickname, setNickname] = useState(''); // Store nickname input
-  const [email, setEmail] = useState(''); // Store email input
-  const [loginUserId, setLoginUserId] = useState(''); // User ID for login input
-  const [loginPassword, setLoginPassword] = useState(''); // Password for login input
   const navigate = useNavigate();
 
   const handleLoginClick = () => {
@@ -22,50 +18,8 @@ const MyPage = () => {
     setView('signup');
   };
 
-  const handlePasswordResetClick = () => {
-    setView('resetPassword');
-  };
-
-  const handleSuccessfulSignUp = () => {
-    if (!userId || !password || !confirmPassword || !nickname || !email) {
-      alert('모든 필드를 입력해 주세요.');
-      return;
-    }
-
-    if (password !== confirmPassword) {
-      alert('비밀번호가 일치하지 않습니다.');
-      return;
-    }
-
-    localStorage.setItem('userData', JSON.stringify({
-      userId,
-      password,
-      nickname,
-      email,
-    }));
-
-    localStorage.setItem('initialUserData', JSON.stringify({
-      userId,
-      password,
-      nickname,
-      email,
-    }));
-
-    setView('success');
-  };
-
   const handleBackToHome = () => {
     navigate('/');
-  };
-
-  const handleLogin = () => {
-    const storedUserData = JSON.parse(localStorage.getItem('userData'));
-
-    if (storedUserData && loginUserId === storedUserData.userId && loginPassword === storedUserData.password) {
-      navigate('/profile', { state: { nickname: storedUserData.nickname || 'User' } });
-    } else {
-      alert('아이디 또는 비밀번호가 일치하지 않습니다.');
-    }
   };
 
   return (
@@ -81,84 +35,9 @@ const MyPage = () => {
         </div>
       )}
 
-      {view === 'login' && (
-        <div className="login-form">
-          <h2>로그인</h2>
-          <input 
-            placeholder="아이디" 
-            value={loginUserId} 
-            onChange={(e) => setLoginUserId(e.target.value)} 
-          />
-          <input 
-            placeholder="비밀번호" 
-            type="password" 
-            value={loginPassword} 
-            onChange={(e) => setLoginPassword(e.target.value)} 
-          />
-          <button onClick={handleLogin}>Login</button>
-          <div className="link-text">
-            <span onClick={handlePasswordResetClick}>비밀번호를 잊어버리셨나요?</span>
-          </div>
-          <div className="link-text">
-            계정이 없으신가요? <span onClick={() => setView('signup')}>sign up</span>
-          </div>
-        </div>
-      )}
-
-      {view === 'signup' && (
-        <div className="signup-form">
-          <h2>회원가입</h2>
-          <input 
-            placeholder="아이디" 
-            value={userId} 
-            onChange={(e) => setUserId(e.target.value)} 
-          />
-          <input 
-            placeholder="비밀번호" 
-            type="password" 
-            value={password} 
-            onChange={(e) => setPassword(e.target.value)} 
-          />
-          <input 
-            placeholder="비밀번호 확인" 
-            type="password" 
-            value={confirmPassword} 
-            onChange={(e) => setConfirmPassword(e.target.value)} 
-          />
-          <input
-            placeholder="별명"
-            value={nickname}
-            onChange={(e) => setNickname(e.target.value)} 
-          />
-          <input 
-            placeholder="Email" 
-            value={email} 
-            onChange={(e) => setEmail(e.target.value)} 
-          />
-          <button onClick={handleSuccessfulSignUp}>Sign up</button>
-          <div className="link-text">
-            이미 계정이 있으신가요? <span onClick={() => setView('login')}>Login</span>
-          </div>
-        </div>
-      )}
-
-      {view === 'resetPassword' && (
-        <div className="reset-password-form">
-          <h2>비밀번호 재설정</h2>
-          <p>비밀번호를 재설정하려면 이메일을 입력하세요.</p>
-          <input 
-            placeholder="Email" 
-            value={email} 
-            onChange={(e) => setEmail(e.target.value)} 
-          />
-          <button onClick={() => alert('비밀번호 재설정 이메일이 전송되었습니다.')}>
-            send
-          </button>
-          <div className="link-text">
-            <span onClick={() => setView('login')}>로그인으로 돌아가기</span>
-          </div>
-        </div>
-      )}
+      {view === 'login' && <Login setView={setView} />}
+      {view === 'signup' && <SignUp setView={setView} />}
+      {view === 'resetPassword' && <ResetPassword setView={setView} />}
 
       {view === 'success' && (
         <div className="success-message">
@@ -174,6 +53,7 @@ const MyPage = () => {
 };
 
 export default MyPage;
+
