+++ src/pages/MyPage/MyPage.scss
@@ -1,3 +1,4 @@
+// src/pages/Home/MyPage.scss
 .mypage-container {
   display: flex;
   align-items: center;
@@ -41,152 +42,54 @@
       cursor: pointer;
     }
   }
+}
 
-  .login-form, .signup-form, .reset-password-form {
-    width: 100%;
-    max-width: 375px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: transparent;
-    padding: 0 43px;
+// src/pages/Home/MyPage.scss (continued)
+.success-message {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+  background-color: #977052;
+  color: #5c3d2e;
+  padding: 0;
+  font-size: 1.5rem;
+  font-weight: bold;
+  text-align: center;
+  position: fixed;
+  top: 0;
+  left: 0;
 
-    h2 {
-      font-size: 24px;
-      color: #ffffff;
-      margin-bottom: 10px;
-    }
-
-    input {
-      width: 100%;
-      padding: 8px;
-      margin: 4px 0;
-      border: none;
-      border-bottom: 1px solid #ffffff;
-      background-color: transparent;
-      color: #ffffff;
-      font-size: 16px;
-
-      &::placeholder {
-        color: #ffffff;
-      }
-    }
-
-    button {
-      width: 100%;
-      padding: 12px;
-      margin-top: 20px;
-      border-radius: 24px;
-      border: none;
-      background-color: #f5c265;
-      color: #5e3a25;
-      font-size: 18px;
-      font-weight: bold;
-      cursor: pointer;
-    }
-
-    .link-text {
-      margin-top: 16px;
-      font-size: 14px;
-      color: #ffffff;
-
-      span {
-        color: #f5c265;
-        cursor: pointer;
-        text-decoration: underline;
-      }
-    }
+  h2 {
+    font-size: 1.8rem;
+    margin-bottom: 20px;
   }
 
-  .reset-password-form {
-    h2 {
-      font-size: 24px;
-      color: #ffffff;
-      margin-bottom: 8px;
-    }
-
-    p {
-      color: #ffffff;
-      font-size: 14px;
-      margin-bottom: 20px;
-      text-align: center;
-    }
-
-    input {
-      width: 100%;
-      padding: 8px;
-      margin: 10px 0 20px;
-      border: none;
-      border-bottom: 1px solid #ffffff;
-      background-color: transparent;
-      color: #ffffff;
-      font-size: 16px;
-      text-align: center;
-
-      &::placeholder {
-        color: #ffffff;
-      }
-    }
+  .button-group {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+    margin-top: 15px;
+    width: 100%;
 
     button {
-      width: 100%;
+      width: 70%;
+      max-width: 250px;
       padding: 12px;
       border-radius: 24px;
       border: none;
       background-color: #f5c265;
-      color: #5e3a25;
-      font-size: 18px;
+      color: #5c3d2e;
+      font-size: 1rem;
       font-weight: bold;
       cursor: pointer;
-      text-transform: capitalize;
-    }
-  }
-
-  .success-message {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-    background-color: #fdfdfd;
-    color: #5c3d2e;
-    padding: 0;
-    font-size: 1.5rem;
-    font-weight: bold;
-    text-align: center;
-    position: fixed;
-    top: 0;
-    left: 0;
-
-    h2 {
-      font-size: 1.8rem;
-      margin-bottom: 20px;
-    }
-
-    .button-group {
-      display: flex;
-      flex-direction: column;
-      gap: 8px;
-      margin-top: 15px;
-      width: 100%;
-
-      button {
-        width: 70%;
-        max-width: 250px;
-        padding: 12px;
-        border-radius: 24px;
-        border: none;
-        background-color: #f5c265;
-        color: #5c3d2e;
-        font-size: 1rem;
-        font-weight: bold;
-        cursor: pointer;
-        align-self: center;
-      }
+      align-self: center;
     }
   }
 }
+
