+++ src/pages/MyPage/ResetPassword.scss
@@ -1,0 +1,53 @@
+// src/pages/Home/ResetPassword.scss
+.reset-password-form {
+    width: 100%;
+    max-width: 375px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: transparent;
+    padding: 0 43px;
+  
+    h2 {
+      font-size: 24px;
+      color: #ffffff;
+      margin-bottom: 8px;
+    }
+  
+    p {
+      color: #ffffff;
+      font-size: 14px;
+      margin-bottom: 20px;
+      text-align: center;
+    }
+  
+    input {
+      width: 100%;
+      padding: 8px;
+      margin: 10px 0 20px;
+      border: none;
+      border-bottom: 1px solid #ffffff;
+      background-color: transparent;
+      color: #ffffff;
+      font-size: 16px;
+      text-align: center;
+  
+      &::placeholder {
+        color: #ffffff;
+      }
+    }
+  
+    button {
+      width: 100%;
+      padding: 12px;
+      border-radius: 24px;
+      border: none;
+      background-color: #f5c265;
+      color: #5e3a25;
+      font-size: 18px;
+      font-weight: bold;
+      cursor: pointer;
+      text-transform: capitalize;
+    }
+  }
+
